add unit tests
Summary:
Add 3 unit tests for SummaryStatistics
(1) The SummaryStatistics should return identical results to what the off-the-shelf functions return, when apply to the non-noisy data.
(2) The SummaryStatistics should debias the summary statistics when applied to noisy data. The results should be similar to what the off-the-shelf functions provide when applied to the underlying non-noisy data.
(3) The DatabaseProcessor and DataFrameProcessor should return the same intermediate results for SummaryStatistics, when applied the same training data. It will further ensure the same final results.

Reviewed By: bkarrer

Differential Revision: D23393549

fbshipit-source-id: dc8442af90e81e619468045f1678c90e5e856a27