unified the code for the dataframe version and the database version
Summary:
There are two versions of the model, one for the in-memory data and the one for data base data.

Before: the in-memory version and the database vesion have their own implementation of sufficient statistics etc. The database version is usually more complicated and tedious than the in-memory version.

After: this diff makes it possible that the database version can be coded the same as the in-memory version. It will greatly reduce the complexity when implementating the score and jacobian for the database version of the logistic regression.

Reviewed By: wierzba3

Differential Revision: D25527017

fbshipit-source-id: 0879d11f40d8d84da8fcab671698439d1f9936ec