update athena procesor
Summary:
(1) add the filters option so that the user can specify filter conditions as input parameter instead of creating a subset of the training data.
(2) add unit test for the filter condition. The testing logic is that the intermediate results that is computed via database version with the full training dataset and filter conditions shold be the same as that computed via dataframe version with subset of the dataset using the same filtering.

Differential Revision: D25639879

fbshipit-source-id: d1c59d0240384e134ac68f500d37e8b9071d583b